Fixed ChordDetailView chord name display, renamed Degree.short to numeric and tweaked some values; added keyboard, note name and display to ChordDetailView

+++ ChordCombiner-SwiftUI/Models/Music Theory Models/Notes/Note.swift
@@ -22,7 +22,7 @@ struct Note: GettableKeyName, Enharmonic, KeySwitch, CustomStringConvertible {
   
   /// degree of the chord or scale
   var degreeName: (name: String, short: String, long: String) {
-    (name: degree.name, short: degree.short, long: degree.long)
+    (name: degree.name, short: degree.numeric, long: degree.long)
   }
   
   var degree: Degree
